Protect some API endpoints using auth.js

+++ app/api/issues/route.ts
@@ -4,8 +4,12 @@ import { NextRequest, NextResponse } from 'next/server';
 import { formatErrors } from '../_utils/format-errors';
 import slugify from 'slugify';
 import { nanoid } from 'nanoid';
+import { auth } from '@/auth';
 
 export async function POST(request: NextRequest) {
+  const session = await auth();
+  if (!session) return NextResponse.json({}, { status: 401 });
+
   const body = await request.json();
   const validationResult = createIssueSchema.safeParse(body);
   if (!validationResult.success)
@@ -22,6 +26,9 @@ export async function GET() {
 }
 
 export async function DELETE(request: NextRequest) {
+  const session = await auth();
+  if (!session) return NextResponse.json({}, { status: 401 });
+
   const deletedIssues = await prisma.issue.deleteMany();
   return NextResponse.json(deletedIssues);
 }
